Start on proof for attitude controller
Adding the attitude controller and its derivation. Much of the work
is from previous papers/books but we can think about just including
the constrained attitude control work. The attitude controller is the
same but we can also avoid pointing at things which might be useful.

The only issue is that all of teh code and examples need to be redone.